move prompt stuff to different file
