+++ groundstation/gs_shell.py
@@ -9,93 +9,11 @@ from utils import read_loop
 import radio_utils.commands as commands
 from radio_utils.disk_buffered_message import DiskBufferedMessage
 from radio_utils import headers
-
 from lib.pycubed_rfm9x_fsk import RFM9x
-
-# print formatters
-bold = '\033[1m'
-normal = '\033[0m'
-red = '\033[31m'
-green = '\033[32m'
-yellow = '\033[33m'
-blue = '\033[34m'
+from shell_utils import bold, normal, red, green, yellow, blue, get_input_discrete, get_input_range
 
 
-def get_input_discrete(prompt_str, choice_values):
-    print(prompt_str)
-    choice = None
-
-    choice_values_str = "("
-    for i, _ in enumerate(choice_values):
-        choice_values_str += f"{choice_values[i]}"
-        if i < len(choice_values) - 1:
-            choice_values_str += ", "
-    choice_values_str += ")"
-
-    choice_values = [cv.lower() for cv in choice_values]
-
-    while choice not in choice_values:
-        choice = input(f"{choice_values_str} ~> ").lower()
-    return choice
-
-
-def set_param_from_input_discrete(param, prompt_str, choice_values, allow_default=False, type=int):
-
-    # add "enter" as a choice
-    choice_values = [""] + choice_values if allow_default else choice_values
-    prompt_str = prompt_str + \
-        " (enter to skip):" if allow_default else prompt_str
-
-    choice = get_input_discrete(prompt_str, choice_values)
-
-    if choice == "":
-        return param
-    else:
-        return type(choice)
-
-
-def is_number(s):
-    try:
-        float(s)
-        return True
-    except ValueError:
-        return False
-
-
-def get_input_range(prompt_str, choice_range):
-    print(prompt_str)
-    choice = None
-
-    choice_range_str = f"({choice_range[0]} - {choice_range[1]})"
-
-    while True:
-        choice = input(f"{choice_range_str} ~> ").lower()
-        if choice == "":
-            break
-
-        if not is_number(choice):
-            continue
-
-        if float(choice) >= choice_range[0] and float(choice) <= choice_range[1]:
-            break
-    return choice
-
-
-def set_param_from_input_range(param, prompt_str, choice_range, allow_default=False):
-
-    # add "enter" as a choice
-    prompt_str = prompt_str + \
-        " (enter to skip):" if allow_default else prompt_str
-
-    choice = get_input_range(prompt_str, choice_range)
-
-    if choice == "":
-        return param
-    else:
-        return float(choice)
-
-
-print(f"\n{bold}{yellow}Radio GS Terminal{normal}\n")
+print(f"\n{bold}{yellow}PyCubed-Mini Groundstation Shell{normal}\n")
 
 board_str = get_input_discrete(
     f"Select the board {bold}(s){normal}atellite, {bold}(f){normal}eather, {bold}(r){normal}aspberry pi",
