+++ groundstation/shell_utils.py
@@ -1,0 +1,84 @@
+"""
+A collection of functions for getting and validating inputs
+"""
+
+# print formatters
+bold = '\033[1m'
+normal = '\033[0m'
+red = '\033[31m'
+green = '\033[32m'
+yellow = '\033[33m'
+blue = '\033[34m'
+
+def get_input_discrete(prompt_str, choice_values):
+    print(prompt_str)
+    choice = None
+
+    choice_values_str = "("
+    for i, _ in enumerate(choice_values):
+        choice_values_str += f"{choice_values[i]}"
+        if i < len(choice_values) - 1:
+            choice_values_str += ", "
+    choice_values_str += ")"
+
+    choice_values = [cv.lower() for cv in choice_values]
+
+    while choice not in choice_values:
+        choice = input(f"{choice_values_str} ~> ").lower()
+    return choice
+
+
+def set_param_from_input_discrete(param, prompt_str, choice_values, allow_default=False, type=int):
+
+    # add "enter" as a choice
+    choice_values = [""] + choice_values if allow_default else choice_values
+    prompt_str = prompt_str + \
+        " (enter to skip):" if allow_default else prompt_str
+
+    choice = get_input_discrete(prompt_str, choice_values)
+
+    if choice == "":
+        return param
+    else:
+        return type(choice)
+
+
+def is_number(s):
+    try:
+        float(s)
+        return True
+    except ValueError:
+        return False
+
+
+def get_input_range(prompt_str, choice_range):
+    print(prompt_str)
+    choice = None
+
+    choice_range_str = f"({choice_range[0]} - {choice_range[1]})"
+
+    while True:
+        choice = input(f"{choice_range_str} ~> ").lower()
+        if choice == "":
+            break
+
+        if not is_number(choice):
+            continue
+
+        if float(choice) >= choice_range[0] and float(choice) <= choice_range[1]:
+            break
+    return choice
+
+
+def set_param_from_input_range(param, prompt_str, choice_range, allow_default=False):
+
+    # add "enter" as a choice
+    prompt_str = prompt_str + \
+        " (enter to skip):" if allow_default else prompt_str
+
+    choice = get_input_range(prompt_str, choice_range)
+
+    if choice == "":
+        return param
+    else:
+        return float(choice)
